requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sentence-transformers==2.3.1
pinecone-client==5.0.1
PyPDF2==3.0.1
python-dotenv==1.0.0


